+++ src/components/03-organisms/Agenda/styles.scss
@@ -1,27 +1,10 @@
-.user {
-  display: flex;
-  justify-content: flex-end;
-  padding: 2rem;
-  margin-bottom: 3rem;
-}
-
-.name {
-  color: white;
-  font-weight: 100;
-}
-
-.hello {
-  color: #faff29;
-  font-weight: 600;
-  margin-right: 5pt;
-}
-
 .agenda {
   font-weight: inherit;
   position: relative;
 }
 
-.container {
-  display: flex;
-  margin: 3rem;
+.loading {
+  color: white;
+  border: 10px solid pink;
+  padding: 3rem;
 }
