--- src/components/App/styles.scss
@@ -1,9 +1,0 @@
-@import '../../shared-styles/bookit';
-
-.content {
-  background-color: $darkblue;
-  margin: 0;
-  padding: 0;
-  height: 100vh;
-}
-
