+++ src/containers/App/styles.scss
@@ -1,0 +1,30 @@
+@import '../../shared-styles/bookit';
+
+.app {
+  background-color: $darkblue;
+  margin: 0;
+  padding: 0;
+  height: 100vh;
+}
+
+.user {
+  display: flex;
+  justify-content: flex-end;
+  padding: 2rem;
+  margin-bottom: 3rem;
+}
+
+.name {
+  color: white;
+  font-weight: 100;
+}
+
+.hello {
+  color: #faff29;
+  font-weight: 600;
+  margin-right: 5pt;
+}
+
+.container {
+  display: flex;
+}
